Add Reporter tests
