--- AIPscan/Data/tests/conftest.py
@@ -1,293 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""This module defines shared Data blueprint pytest fixtures."""
-
-from datetime import datetime
-
-import pytest
-
-from AIPscan import create_app, db, test_helpers
-from AIPscan.models import FileType
-
-TIFF_FILE_FORMAT = "Tagged Image File Format"
-TIFF_PUID = "fmt/353"
-
-JPEG_FILE_FORMAT = "JPEG"
-
-JPEG_1_01_FORMAT_VERSION = "1.01"
-JPEG_1_01_PUID = "fmt/43"
-
-JPEG_1_02_FORMAT_VERSION = "1.02"
-JPEG_1_02_PUID = "fmt/44"
-
-ORIGINAL_FILE_SIZE = 1000
-PRESERVATION_FILE_SIZE = 2000
-
-AIP_1_CREATION_DATE = "2020-01-01"
-AIP_2_CREATION_DATE = "2020-06-01"
-
-AIP_1_NAME = "TestAIP1"
-AIP_2_NAME = "TestAIP2"
-
-AIP_1_UUID = "111111111111-1111-1111-11111111"
-AIP_2_UUID = "222222222222-2222-2222-22222222"
-
-ORIGINAL_FILE_1_NAME = "original-file-1.jpg"
-ORIGINAL_FILE_1_UUID = "333333333333-3333-3333-33333333"
-
-ORIGINAL_FILE_2_NAME = "original-file-2.jpg"
-ORIGINAL_FILE_2_UUID = "444444444444-4444-4444-44444444"
-
-PRESERVATION_FILE_1_NAME = "preservation-file-1.tif"
-PRESERVATION_FILE_2_NAME = "preservation-file-2.tif"
-
-PRESERVATION_FILE_1_UUID = "555555555555-5555-5555-55555555"
-PRESERVATION_FILE_2_UUID = "666666666666-6666-6666-66666666"
-
-STORAGE_SERVICE_NAME = "Test Storage Service"
-
-
-def _datetime_obj_from_string(date_string, date_only=False):
-    """Return a datetime object from a given string to aid in the
-    assignment of consistent and reliable test dates.
-    """
-    str_format = "%Y-%m-%d %H:%M:%S"
-    if date_only is True:
-        str_format = "%Y-%m-%d"
-
-    return datetime.strptime(date_string, str_format)
-
-
-INGEST_EVENT_CREATION_TIME = _datetime_obj_from_string("2020-12-02 10:00:00")
-AIP_CREATION_TIME = _datetime_obj_from_string("2020-12-02 10:30:32")
-
-
-@pytest.fixture
-def app_with_populated_files(scope="package"):
-    """Fixture with pre-populated data.
-
-    This fixture is used to create expected state which is then used to
-    test the Data.aips_by_file_format and Data.aips_by_puid endpoints.
-    """
-    app = create_app("test")
-    with app.app_context():
-        db.create_all()
-
-        storage_service = test_helpers.create_test_storage_service()
-        fetch_job = test_helpers.create_test_fetch_job(
-            storage_service_id=storage_service.id
-        )
-
-        _ = test_helpers.create_test_aip(
-            uuid=AIP_1_UUID,
-            create_date=AIP_CREATION_TIME,
-            storage_service_id=storage_service.id,
-            fetch_job_id=fetch_job.id,
-        )
-
-        original_file = test_helpers.create_test_file(
-            file_type=FileType.original,
-            size=ORIGINAL_FILE_SIZE,
-            puid=TIFF_PUID,
-            file_format=TIFF_FILE_FORMAT,
-        )
-
-        _ = test_helpers.create_test_file(
-            file_type=FileType.preservation,
-            size=PRESERVATION_FILE_SIZE,
-            puid=TIFF_PUID,
-            file_format=TIFF_FILE_FORMAT,
-        )
-
-        user_agent = test_helpers.create_test_agent()
-
-        event = test_helpers.create_test_event(
-            type="ingestion", date=INGEST_EVENT_CREATION_TIME, file_id=original_file.id
-        )
-
-        _ = test_helpers.create_test_event_agent(
-            event_id=event.id, agent_id=user_agent.id
-        )
-
-        yield app
-
-        db.drop_all()
-
-
-@pytest.fixture
-def app_with_populated_files_no_ingestion_event(scope="package"):
-    """Fixture with pre-populated data.
-
-    This fixture is used to create expected state which is then used to
-    test the Data.aips_by_file_format and Data.aips_by_puid endpoints.
-    """
-    app = create_app("test")
-    with app.app_context():
-        db.create_all()
-
-        storage_service = test_helpers.create_test_storage_service()
-        fetch_job = test_helpers.create_test_fetch_job(
-            storage_service_id=storage_service.id
-        )
-
-        _ = test_helpers.create_test_aip(
-            uuid=AIP_1_UUID,
-            create_date=AIP_CREATION_TIME,
-            storage_service_id=storage_service.id,
-            fetch_job_id=fetch_job.id,
-        )
-
-        test_helpers.create_test_file(
-            file_type=FileType.original,
-            size=ORIGINAL_FILE_SIZE,
-            puid=TIFF_PUID,
-            file_format=TIFF_FILE_FORMAT,
-        )
-
-        _ = test_helpers.create_test_file(
-            file_type=FileType.preservation,
-            size=PRESERVATION_FILE_SIZE,
-            puid=TIFF_PUID,
-            file_format=TIFF_FILE_FORMAT,
-        )
-
-        yield app
-
-        db.drop_all()
-
-
-@pytest.fixture
-def app_with_populated_format_versions(scope="package"):
-    """Fixture with pre-populated data.
-
-    This fixture is used to create expected state which is then used to
-    test the Data.format_versions_count endpoint.
-    """
-    AIP_DATE_FORMAT = "%Y-%m-%d"
-
-    app = create_app("test")
-    with app.app_context():
-        db.create_all()
-
-        storage_service = test_helpers.create_test_storage_service()
-        fetch_job = test_helpers.create_test_fetch_job(
-            storage_service_id=storage_service.id
-        )
-
-        aip1 = test_helpers.create_test_aip(
-            create_date=datetime.strptime(AIP_1_CREATION_DATE, AIP_DATE_FORMAT),
-            storage_service_id=storage_service.id,
-            fetch_job_id=fetch_job.id,
-        )
-
-        aip2 = test_helpers.create_test_aip(
-            create_date=datetime.strptime(AIP_2_CREATION_DATE, AIP_DATE_FORMAT),
-            storage_service_id=storage_service.id,
-            fetch_job_id=fetch_job.id,
-        )
-
-        _ = test_helpers.create_test_file(
-            size=ORIGINAL_FILE_SIZE,
-            puid=JPEG_1_01_PUID,
-            file_format=JPEG_FILE_FORMAT,
-            format_version=JPEG_1_01_FORMAT_VERSION,
-            aip_id=aip1.id,
-        )
-
-        _ = test_helpers.create_test_file(
-            size=PRESERVATION_FILE_SIZE,
-            puid=JPEG_1_02_PUID,
-            file_format=JPEG_FILE_FORMAT,
-            format_version=JPEG_1_02_FORMAT_VERSION,
-            aip_id=aip2.id,
-        )
-
-        _ = test_helpers.create_test_file(
-            size=None,
-            puid="fmt/468",
-            file_format="ISO Disk Image File",
-            format_version=None,
-            aip_id=aip2.id,
-        )
-
-        yield app
-
-        db.drop_all()
-
-
-@pytest.fixture
-def app_with_populated_preservation_derivatives(scope="package"):
-    """Fixture with pre-populated data.
-
-    This fixture is used to create expected state which is then used to
-    test the report_data.preservation_derivatives endpoint.
-    """
-    app = create_app("test")
-    with app.app_context():
-        db.create_all()
-
-        storage_service = test_helpers.create_test_storage_service(
-            name=STORAGE_SERVICE_NAME
-        )
-        fetch_job = test_helpers.create_test_fetch_job(
-            storage_service_id=storage_service.id
-        )
-
-        aip1 = test_helpers.create_test_aip(
-            uuid=AIP_1_UUID,
-            transfer_name=AIP_1_NAME,
-            storage_service_id=storage_service.id,
-            fetch_job_id=fetch_job.id,
-        )
-        aip2 = test_helpers.create_test_aip(
-            uuid=AIP_2_UUID,
-            transfer_name=AIP_2_NAME,
-            storage_service_id=storage_service.id,
-            fetch_job_id=fetch_job.id,
-        )
-
-        original_file1 = test_helpers.create_test_file(
-            file_type=FileType.original,
-            name=ORIGINAL_FILE_1_NAME,
-            uuid=ORIGINAL_FILE_1_UUID,
-            size=ORIGINAL_FILE_SIZE,
-            puid=JPEG_1_01_PUID,
-            file_format=JPEG_FILE_FORMAT,
-            format_version=JPEG_1_01_FORMAT_VERSION,
-            aip_id=aip1.id,
-        )
-        original_file2 = test_helpers.create_test_file(
-            file_type=FileType.original,
-            name=ORIGINAL_FILE_2_NAME,
-            uuid=ORIGINAL_FILE_2_UUID,
-            size=ORIGINAL_FILE_SIZE,
-            puid=JPEG_1_02_PUID,
-            file_format=JPEG_FILE_FORMAT,
-            format_version=JPEG_1_02_FORMAT_VERSION,
-            aip_id=aip2.id,
-        )
-
-        _ = test_helpers.create_test_file(
-            file_type=FileType.preservation,
-            name=PRESERVATION_FILE_1_NAME,
-            uuid=PRESERVATION_FILE_1_UUID,
-            size=PRESERVATION_FILE_SIZE,
-            puid=TIFF_PUID,
-            file_format=TIFF_FILE_FORMAT,
-            original_file_id=original_file1.id,
-            aip_id=aip1.id,
-        )
-        _ = test_helpers.create_test_file(
-            file_type=FileType.preservation,
-            name=PRESERVATION_FILE_2_NAME,
-            uuid=PRESERVATION_FILE_2_UUID,
-            size=PRESERVATION_FILE_SIZE,
-            puid=TIFF_PUID,
-            file_format=TIFF_FILE_FORMAT,
-            original_file_id=original_file2.id,
-            aip_id=aip2.id,
-        )
-
-        yield app
-
-        db.drop_all()
