+++ AIPscan/Data/tests/test_agents_transfers.py
@@ -6,8 +6,8 @@ the ingest log report and agents-transfers API endpoint.
 
 import pytest
 
+from AIPscan.conftest import AIP_CREATION_TIME, INGEST_EVENT_CREATION_TIME
 from AIPscan.Data import report_data
-from AIPscan.Data.tests.conftest import AIP_CREATION_TIME, INGEST_EVENT_CREATION_TIME
 
 
 @pytest.mark.parametrize(
