+++ AIPscan/Data/tests/test_aips_by_format.py
@@ -2,6 +2,11 @@
 
 import pytest
 
+from AIPscan.conftest import (
+    ORIGINAL_FILE_SIZE,
+    PRESERVATION_FILE_SIZE,
+    TIFF_FILE_FORMAT,
+)
 from AIPscan.Data import fields, report_data
 from AIPscan.Data.tests import (
     MOCK_AIPS_BY_FORMAT_OR_PUID_QUERY_RESULTS as MOCK_QUERY_RESULTS,
@@ -10,11 +15,6 @@ from AIPscan.Data.tests import (
     MOCK_STORAGE_SERVICE,
     MOCK_STORAGE_SERVICE_ID,
     MOCK_STORAGE_SERVICE_NAME,
-)
-from AIPscan.Data.tests.conftest import (
-    ORIGINAL_FILE_SIZE,
-    PRESERVATION_FILE_SIZE,
-    TIFF_FILE_FORMAT,
 )
