+++ AIPscan/Reporter/tests/test_aips_by_puid.py
@@ -1,12 +1,18 @@
-# -*- coding: utf-8 -*-
-
 import datetime
 import uuid
 
 import pytest
+from flask import current_app
 
 from AIPscan.models import File, FileType
 from AIPscan.Reporter.report_aips_by_puid import get_format_string_from_puid
+
+EXPECTED_CSV_ORIGINAL = (
+    b"AIP Name,UUID,Count,Size\r\nTest AIP,111111111111-1111-1111-11111111,1,1.0 kB\r\n"
+)
+EXPECTED_CSV_PRESERVATION = (
+    b"AIP Name,UUID,Count,Size\r\nTest AIP,111111111111-1111-1111-11111111,1,2.0 kB\r\n"
+)
 
 FILE_WITH_FORMAT_ONLY = File(
     uuid=uuid.uuid4(),
@@ -74,3 +80,49 @@ def test_get_format_name_and_version_from_puid(
     mock_get_file = mocker.patch("sqlalchemy.orm.query.Query.first")
     mock_get_file.return_value = mock_file
     assert expected_return_value == get_format_string_from_puid(puid)
+
+
+@pytest.mark.parametrize(
+    "original_files",
+    [
+        # Original files.
+        ("True"),
+        # Preservation files.
+        ("False"),
+    ],
+)
+def test_aips_by_puid(app_with_populated_files, original_files):
+    """Test that report template renders."""
+    with current_app.test_client() as test_client:
+        response = test_client.get(
+            "/reporter/aips_by_puid/?amss_id=1&puid=fmt/353&original_files={}".format(
+                original_files
+            )
+        )
+        assert response.status_code == 200
+
+
+@pytest.mark.parametrize(
+    "original_files,expected_csv",
+    [
+        # Original files.
+        ("True", EXPECTED_CSV_ORIGINAL),
+        # Preservation files.
+        ("False", EXPECTED_CSV_PRESERVATION),
+    ],
+)
+def test_aips_by_puid_csv(app_with_populated_files, original_files, expected_csv):
+    """Test CSV export."""
+    with current_app.test_client() as test_client:
+        response = test_client.get(
+            "/reporter/aips_by_puid/?amss_id=1&puid=fmt/353&original_files={}&csv=True".format(
+                original_files
+            )
+        )
+        assert response.status_code == 200
+        assert (
+            response.headers["Content-Disposition"]
+            == "attachment; filename=aips_by_puid_fmt/353.csv"
+        )
+        assert response.mimetype == "text/csv"
+        assert response.data == expected_csv
